+++ app/src/main/java/com/fpf/smartscan/ui/action/SearchAction.kt
@@ -7,8 +7,8 @@ import com.fpf.smartscan.media.MediaItem
 import com.fpf.smartscan.media.MediaType
 
 sealed interface SearchAction {
-    data class Search(val similarityThreshold: Float, val dedupeEnabled: Boolean): SearchAction
-    data class SetQueryImageAndSearch(val image: Uri, val similarityThreshold: Float, val dedupeEnabled: Boolean): SearchAction
+    data class Search(val strictness: Float, val dedupeEnabled: Boolean): SearchAction
+    data class SetQueryImageAndSearch(val image: Uri, val strictness: Float, val dedupeEnabled: Boolean): SearchAction
     data class ViewResult(val context: Context, val item: MediaItem, val autoOpenInGallery: Boolean? = null): SearchAction
     data class ToggleSelectedResult(val item: MediaItem): SearchAction
     data class TagItems(val tag: String): SearchAction